WIP Add new gnl no leak found

+++ includes/libft/libft.h
@@ -38,7 +38,7 @@ int				ft_isalnum(int c);
 int				ft_isalpha(int c);
 int				ft_memcmp(const void *s1, const void *s2, size_t n);
 char			*ft_strchr(const char *s, int c);
-int				ft_strlen(char *c);
+size_t			ft_strlen(char *c);
 char			*ft_strdup(char *src);
 unsigned int	ft_strlcat(char *dest, char *src, unsigned int size);
 size_t			ft_strlcpy(char *dest, const char *src, size_t size);
